+++ core/src/main/java/edu/ucr/cs/riple/core/metadata/field/FieldDeclarationAnalysis.java
@@ -11,6 +11,7 @@ import edu.ucr.cs.riple.core.ModuleInfo;
 import edu.ucr.cs.riple.core.metadata.MetaData;
 import edu.ucr.cs.riple.injector.Helper;
 import edu.ucr.cs.riple.injector.exceptions.TargetClassNotFound;
+import edu.ucr.cs.riple.scanner.TypeAnnotatorScanner;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.Collections;
@@ -30,7 +31,7 @@ public class FieldDeclarationAnalysis extends MetaData<FieldDeclarationInfo> {
   /**
    * Output file name. It contains information about all classes existing in source code. Each line
    * has the format: (class-flat-name \t uri to file containing the class). It is generated by
-   * {@link edu.ucr.cs.riple.scanner.Scanner} checker.
+   * {@link TypeAnnotatorScanner} checker.
    */
   public static final String FILE_NAME = "class_info.tsv";
