+++ core/src/test/resources/templates/downstream-dependency-test/DepC/build.gradle
@@ -42,7 +42,7 @@ dependencies {
     compileOnly project(":Target")
 
     annotationProcessor "com.uber.nullaway:nullaway:0.9.9"
-    annotationProcessor "edu.ucr.cs.riple.nullawayannotator:scanner:1.3.2-SNAPSHOT"
+    annotationProcessor "edu.ucr.cs.riple.nullawayannotator:type-annotator-scanner:1.3.2-SNAPSHOT"
     annotationProcessor files(libraryloader)
 
     compileOnly "com.google.code.findbugs:jsr305:3.0.2"
@@ -61,8 +61,8 @@ tasks.withType(JavaCompile) {
                                               "-XepOpt:NullAway:AnnotatedPackages=test",
                                               "-XepOpt:NullAway:SerializeFixMetadata=true",
                                               "-XepOpt:NullAway:FixSerializationConfigPath=" + project.getProperty("DepC-nullaway-config-path"),
-                                              "-Xep:Scanner:ERROR",
-                                              "-XepOpt:Scanner:ConfigPath=" + project.getProperty("DepC-scanner-config-path"),
+                                              "-Xep:TypeAnnotatorScanner:ERROR",
+                                              "-XepOpt:TypeAnnotatorScanner:ConfigPath=" + project.getProperty("DepC-scanner-config-path"),
                                               "-XepOpt:NullAway:AcknowledgeLibraryModelsOfAnnotatedCode=true",
                                               "-XepDisableAllWarnings"]
         options.compilerArgs << "-Xmaxerrs" << "100000"
