+++ runner/update.sh
@@ -1,2 +1,10 @@
-cd ../ && ./gradlew :scanner:publishToMavenLocal --rerun-tasks && ./gradlew :core:shadowJar --rerun-tasks
-mv core/build/libs/core-1.3.1-SNAPSHOT-all.jar runner/jars/core.jar
+set -exu
+
+CURRENT_VERSION="core-1.3.2-SNAPSHOT-all.jar"
+PROJECT_ROOT=${PROJECT_ROOT:-$(git rev-parse --show-toplevel)}
+
+pushd "$PROJECT_ROOT"
+   ./gradlew :type-annotator-scanner:publishToMavenLocal --rerun-tasks
+   ./gradlew :core:shadowJar --rerun-tasks
+   mv core/build/libs/"$CURRENT_VERSION" runner/jars/core.jar
+popd
