+++ settings.gradle
@@ -24,7 +24,7 @@
 
 rootProject.name = 'NullAwayAnnotator'
 include 'core'
-include 'scanner'
+include 'type-annotator-scanner'
 include 'injector'
 include 'library-model-loader'
 include 'checks'
