+++ type-annotator-scanner/README.md
@@ -1,0 +1,100 @@
+## Type Annotator Scanner (Scanner)
+
+`Type Annotator Scanner` is an Error Prone checker which serializes the required information regarding the source code structure for the Annotator.
+
+Below are the series of information serialized by this checker:
+1. `Region` info where an element in the source code is used.
+2. Method info to construct the `Method Inheritance Tree`.
+3. Class info.
+
+See sections below where each series is described in details
+
+#### Region
+
+An __element__ in the source code is __used__ if:
+- __element__ is a method and the method is invoked.
+- __element__ is a class field and the field is read/initialized.
+
+
+A __Region__ is the enclosing method body where an element is used. Each region is denoted as below:
+1. `method`: Enclosing method signature in `string`.
+2. `class`: Enclosing class flat name in `string`.
+
+If the program point where the element is used is not enclosed by a method (e.g. class initialization region) the `method` value be the string `"null"`.
+
+#### Method Inheritance Tree
+`Method Inheritance Tree` is a tree of nodes where each node corresponds to a method in the source node. Each node's parent in this structure, is the node which corresponds to method's closest overriding method.
+
+A node corresponding to method `m` contains the information below:
+1. Signature of `m`.
+2. Flat name of the enclosing class of `m`.
+3. Annotations of each parameter.
+4. Annotation of the method return.
+5. Unique id of this node.
+6. Unique id of parent node.
+7. Visibility of `m`.
+8. Return type of `m`.
+
+#### Class info
+Serializes flat name of all observed classes and uri of the file containing the class.
+
+This data is used by the `core` module to detect multiple inline field declarations.
+
+### Error Prone Flags / Config File Format
+
+To activate the `Type Annotator Scanner` checker pass the flag below to the build configuration:
+`-Xep:TypeAnnotatorScanner:ERROR`
+
+`Type Annotator Scanner` requires a config file to serialize the requested information at the designated directory. This config file should be written in `xml` format and the path to this file should be passed via the flag below:
+`-XepOpt:TypeAnnotatorScanner:ConfigPath=path_to_scanner.xml`
+
+
+#### Config format
+The config file is written in `xml` format has the structure below:
+```xml
+<scanner>
+    <method activation="true"/>
+    <field activation="true"/>
+    <call activation="true"/>
+    <class activation="true"/>
+    <path> Path to designated output directory </path>
+</scanner>
+```
+
+The config above will be generated by the `core` module automatically.
+
+
+
+### Installation
+To install the `Type Annotator Scanner` please run the command below:
+```shell
+./gradlew :type-annotator-scanner:publishToMavenLocal
+```
+
+The target project must have the configuration below:
+```groovy
+plugins {
+    id 'java'
+    id("net.ltgt.errorprone") version "0.6"
+}
+
+repositories {
+    mavenLocal()
+}
+
+dependencies {
+    annotationProcessor "edu.ucr.cs.riple.nullawayannotator:type-annotator-scanner:1.3.2-SNAPSHOT"
+
+    compileOnly "com.google.code.findbugs:jsr305:3.0.2"
+    errorprone "com.google.errorprone:error_prone_core:2.3.2"
+    errorproneJavac "com.google.errorprone:javac:9+181-r4173-1"
+}
+
+tasks.withType(JavaCompile) {
+    if (!name.toLowerCase().contains("test")) {
+        options.errorprone.errorproneArgs += ["-Xep:TypeAnnotatorScanner:ERROR",
+                                              "-XepOpt:TypeAnnotatorScanner:ConfigPath=path_to_scanner.xml",
+                                              ]
+    }
+}
+```
