+++ type-annotator-scanner/src/main/java/edu/ucr/cs/riple/scanner/Serializer.java
