+++ type-annotator-scanner/src/test/java/edu/ucr/cs/riple/scanner/ClassInfoTest.java
@@ -32,7 +32,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 @RunWith(JUnit4.class)
-public class ClassInfoTest extends ScannerBaseTest<ClassInfoDisplay> {
+public class ClassInfoTest extends TypeAnnotatorScannerBaseTest<ClassInfoDisplay> {
 
   private static final DisplayFactory<ClassInfoDisplay> CLASS_DISPLAY_FACTORY =
       values -> {
