+++ type-annotator-scanner/src/test/java/edu/ucr/cs/riple/scanner/MethodInfoTest.java
@@ -32,7 +32,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 @RunWith(JUnit4.class)
-public class MethodInfoTest extends ScannerBaseTest<MethodInfoDisplay> {
+public class MethodInfoTest extends TypeAnnotatorScannerBaseTest<MethodInfoDisplay> {
 
   private static final DisplayFactory<MethodInfoDisplay> METHOD_DISPLAY_FACTORY =
       values -> {
