+++ type-annotator-scanner/src/test/java/edu/ucr/cs/riple/scanner/tools/ClassInfoDisplay.java
