+++ type-annotator-scanner/src/test/java/edu/ucr/cs/riple/scanner/tools/SerializationTestHelper.java
@@ -30,7 +30,7 @@ import static org.junit.Assert.fail;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import com.google.errorprone.CompilationTestHelper;
-import edu.ucr.cs.riple.scanner.Scanner;
+import edu.ucr.cs.riple.scanner.TypeAnnotatorScanner;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.nio.charset.Charset;
@@ -109,14 +109,14 @@ public class SerializationTestHelper<T extends Display> {
   }
 
   /**
-   * Creates the actual {@link Scanner} with the given arguments. This method is part of the builder
-   * pattern and should be called before any other method.
+   * Creates the actual {@link TypeAnnotatorScanner} with the given arguments. This method is part
+   * of the builder pattern and should be called before any other method.
    *
    * @return Receiver of the call.
    */
   public SerializationTestHelper<T> setArgs(List<String> args) {
     compilationTestHelper =
-        CompilationTestHelper.newInstance(Scanner.class, getClass()).setArgs(args);
+        CompilationTestHelper.newInstance(TypeAnnotatorScanner.class, getClass()).setArgs(args);
     return this;
   }
