+++ type-annotator-scanner/src/test/resources/edu/ucr/cs/riple/scanner/SampleClassForTest.java
